ScheduleScreen logic split moved to schedule package
