+++ app/src/main/java/com/nailorsh/repeton/ui/screens/schedule/Day.kt
@@ -1,0 +1,71 @@
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import com.nailorsh.repeton.R
+import java.time.DayOfWeek
+import java.time.LocalDate
+
+@Composable
+fun Day(day: LocalDate, selectedDay: LocalDate, onSelectDay: () -> Unit) {
+    val selected = day.equals(selectedDay)
+    val backgroundColor =
+        if (selected) MaterialTheme.colorScheme.tertiary else MaterialTheme.colorScheme.tertiaryContainer
+    val textColor =
+        if (selected) MaterialTheme.colorScheme.onTertiary else MaterialTheme.colorScheme.onTertiaryContainer
+    val numberTextStyle = MaterialTheme.typography.headlineSmall
+    val dayTextStyle = MaterialTheme.typography.labelSmall
+    val dayFontWeight = FontWeight.W400
+
+    Box(
+        modifier = Modifier
+            .width(36.dp)
+            .height(48.dp)
+            .background(color = backgroundColor, shape = MaterialTheme.shapes.small)
+            .border(width = 1.dp, color = Color.Black, shape = MaterialTheme.shapes.small)
+            .padding(horizontal = 2.dp)
+            .clickable(onClick = onSelectDay)
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxHeight()
+                .fillMaxWidth(),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.spacedBy((-4).dp)
+        ) {
+            Text(
+                text = day.dayOfMonth.toString(),
+                color = textColor,
+                textAlign = TextAlign.Center,
+                style = numberTextStyle,
+                modifier = Modifier.padding(top = 6.dp),
+            )
+            Text(
+                text = when (day.dayOfWeek) {
+                    DayOfWeek.MONDAY -> stringResource(R.string.mon)
+                    DayOfWeek.TUESDAY -> stringResource(R.string.tue)
+                    DayOfWeek.WEDNESDAY -> stringResource(R.string.wed)
+                    DayOfWeek.THURSDAY -> stringResource(R.string.thu)
+                    DayOfWeek.FRIDAY -> stringResource(R.string.fri)
+                    DayOfWeek.SATURDAY -> stringResource(R.string.sat)
+                    DayOfWeek.SUNDAY -> stringResource(R.string.sun)
+                    else -> ""
+                },
+                color = textColor,
+                textAlign = TextAlign.Center,
+                style = dayTextStyle,
+                fontWeight = dayFontWeight
+            )
+        }
+    }
+}
