+++ app/src/main/java/com/nailorsh/repeton/ui/screens/schedule/LessonBox.kt
@@ -1,0 +1,86 @@
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.nailorsh.repeton.R
+import com.nailorsh.repeton.model.Lesson
+import java.time.format.DateTimeFormatter
+import java.time.format.FormatStyle
+
+@Composable
+fun LessonBox(lesson: Lesson, onClick: (Lesson) -> Unit, modifier: Modifier = Modifier) {
+
+    Box(
+        modifier = Modifier
+            .padding(top = 21.dp)
+            .width(dimensionResource(R.dimen.schedule_screen_button_width))
+            .height(dimensionResource(R.dimen.schedule_screen_lesson_size))
+            .background(
+                color = MaterialTheme.colorScheme.secondaryContainer,
+                shape = MaterialTheme.shapes.medium
+            )
+            .clickable { onClick(lesson) }
+    )
+    {
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(horizontal = 13.dp)
+
+        ) {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = 10.dp),
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.SpaceBetween
+            )
+            {
+                Text(
+                    text = lesson.subject,
+                    color = MaterialTheme.colorScheme.onSecondaryContainer,
+                    style = MaterialTheme.typography.headlineSmall,
+
+                    )
+
+
+                // Время начала и конца занятия
+                val startTimeCutted =
+                    lesson.startTime.format(DateTimeFormatter.ofLocalizedDateTime(FormatStyle.SHORT))
+                        .substringAfter(", ")
+                val endTimeCutted =
+                    lesson.endTime.format(DateTimeFormatter.ofLocalizedDateTime(FormatStyle.SHORT))
+                        .substringAfter(", ")
+                Text(
+                    text = "$startTimeCutted - $endTimeCutted",
+                    style = MaterialTheme.typography.bodySmall,
+                    fontWeight = FontWeight.Medium,
+                    textAlign = TextAlign.Left,
+                    modifier = Modifier
+                        .padding(start = 4.dp)
+                )
+            }
+            Text(
+                text = lesson.teacherName,
+                color = MaterialTheme.colorScheme.onSecondaryContainer,
+                style = MaterialTheme.typography.bodyMedium,
+                letterSpacing = 0.25.sp,
+            )
+            Text(
+                text = lesson.title,
+                style = MaterialTheme.typography.labelSmall,
+                color = MaterialTheme.colorScheme.onSecondaryContainer.copy(alpha = 0.6f),
+                fontWeight = FontWeight.SemiBold
+            )
+        }
+    }
+}
